pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitgn-contest-agent"
version = "0.1.28"
description = "BitGN PAC1 contest agent (single-session, provider-agnostic SGR)"
requires-python = ">=3.12"
dependencies = [
  "pydantic>=2.6",
  "openai>=1.40",
  "PyYAML>=6.0",
  # BitGN proto/connectrpc stubs are published via buf.build/gen/python
  # (the same source sample-agents/pac1-py pins). The local-sdk path that
  # used to live here only existed on the Linux dev box.
  "bitgn-api-connectrpc-python==0.9.0.1.20260324104753+6b18c777b19a",
  "bitgn-api-protocolbuffers-python==34.1.0.1.20260324104753+6b18c777b19a",
  "connectrpc",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-mock>=3.12",
]

[project.scripts]
bitgn-agent = "bitgn_contest_agent.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[[tool.uv.index]]
url = "https://buf.build/gen/python"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
